Add registration event.
