+++ src/main/java/ru/andnovikov/sportnow/domain/enumeration/RegStatus.java
@@ -1,0 +1,5 @@
+package ru.andnovikov.sportnow.domain.enumeration;
+
+public enum RegStatus {
+    NEW, PAYED
+}
